+++ src/Adapter/InMemory/Overwrite.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Filesystem\Adapter\InMemory;
+
+use Innmind\Filesystem\{
+    File,
+    Directory,
+};
+
+/**
+ * @internal
+ */
+final class Overwrite
+{
+    public function __invoke(Directory $parent, File $file): Directory
+    {
+        return $parent->add($file);
+    }
+}